+++ modules/yup_python/yup_python_data_model.cpp
@@ -19,4 +19,6 @@
   ==============================================================================
 */
 
+#if YUP_MODULE_AVAILABLE_yup_data_model
 #include "bindings/yup_YupDataModel_bindings.cpp"
+#endif