+++ modules/yup_python/yup_python_graphics.cpp
@@ -19,4 +19,6 @@
   ==============================================================================
 */
 
+#if YUP_MODULE_AVAILABLE_yup_graphics
 #include "bindings/yup_YupGraphics_bindings.cpp"
+#endif